Move /latest-job endpoint to latest_job.py
